Fix issue with multi-step Undo
Previously, I missed rollback of the status.

This implementation saves "route" <source, destination> pair in the
Event. Then before we execute an Event we copy the anti-Event (with
same payload but opposite "route") to the "undo" railway.

Also, I found that we missed the finalize_undo action, when you need to
split changes between internal and external for the Pipeline.

The idea is the same as for pair call -> finalize